+++ nczarr_test/testfilter_multi.c
@@ -43,7 +43,7 @@ static const char* testfile = NULL;
 
 static size_t dimsize = DIMSIZE;
 static size_t chunksize = CHUNKSIZE;
-static size_t actualdims = NDIMS;
+static int actualdims = NDIMS;
 
 static size_t actualproduct = 1; /* x-product over dim sizes */
 static size_t chunkproduct = 1; /* x-product over chunksizes */
