+++ nczarr_test/testfilter_repeat.c
@@ -34,7 +34,7 @@
 static size_t dimsize[NDIMS] = {4,4,4,4};
 static size_t chunksize[NDIMS] = {4,4,4,4};
 
-static size_t ndims = NDIMS;
+static int ndims = NDIMS;
 
 static size_t totalproduct = 1; /* x-product over max dims */
 static size_t actualproduct = 1; /* x-product over actualdims */
@@ -318,8 +318,8 @@ odom_offset(void)
     int i;
     int offset = 0;
     for(i=0;i<ndims;i++) {
-        offset *= dimsize[i];
-        offset += odom[i];
+        offset *= (int)dimsize[i];
+        offset += (int)odom[i];
     }
     return offset;
 }
@@ -331,8 +331,8 @@ expectedvalue(void)
     float offset = 0;
 
     for(i=0;i<ndims;i++) {
-        offset *= dimsize[i];
-        offset += odom[i];
+        offset *= (float)dimsize[i];
+        offset += (float)odom[i];
     }
     return offset;
 }
